+++ src/tools/DeepAnalysis/FnVar/def/OutputVarCommandPseudoArray.ts
@@ -1,0 +1,5 @@
+/**
+ * <https://github.com/CoffeeChaton/vscode-autohotkey-NekoHelp/issues/11>
+ */
+
+const z = 0;